Migrate warp-rust to use sqlx (#6354)
In the past, warp-rust has used tokio-postgres which because
it was compatible with async code and there was no real
alternative to it - it wasn't great, but there was nothing
better.

However, since then sqlx was released which provides a nicer
interface than very low-level interface of tokio-postgres
while providing nice features such as built-in connection
pooling, automatic statement preparation and caching,
fetching results directly into structures, and more.
