+++ frameworks/Rust/warp-rust/README.md
@@ -2,7 +2,7 @@
 
 warp is a composable web server framework based on hyper.
 
-* [API Documentation](https://docs.rs/warp/0.2)
+* [API Documentation](https://docs.rs/warp/0.3)
 
 ### Test Type Implementation Source Code
 
@@ -10,6 +10,7 @@ warp is a composable web server framework based on hyper.
 * [PLAINTEXT](src/main.rs)
 * [DB](src/main.rs)
 * [QUERIES](src/main.rs)
+* [FORTUNES](src/main.rs)
 
 ## Test URLs
 ### JSON
@@ -27,3 +28,7 @@ http://localhost:8080/db
 ### QUERIES
 
 http://localhost:8080/queries/[1...500]
+
+### FORTUNES
+
+http://localhost:8080/fortunes
